A commit change all over the place.
Basically, the two new components expected different information at
different points.

Now, when going to the new route a zitting is created.
From there on the remaining forms can be inserted.

I tried also to make the saving more modular, per component, since a
big fat save button will make the process super slow once the amount
of data starts to grow.

I cleaned some stuff too, and extracted some components.

I am sorry for the messy commit. But time is not my friend.
