+++ app/components/manage-agenda-zitting.js
@@ -4,14 +4,12 @@ import { action } from '@ember/object';
 export default class ManageAgendaZittingComponent extends Component {
   constructor(...args){
     super(...args);
-
-    this.args.zitting.agendapunten=this.args.zitting.agendapunten.sortBy('position');
+    this.args.zitting.agendapunten = this.args.zitting.agendapunten.sortBy('position');
   }
-  @tracked popup=false;
+  @tracked popup = false;
 
   @action
   cancel(){
-    this.popup=false;
+    this.popup = false;
   }
 }
-
