+++ app/components/meeting-form.hbs
@@ -1,41 +1,23 @@
-Meeting form[@TN]
-<div {{did-insert this.fetchBestuursorgaan}}>
-    <div>
-      Planned Start[@TN]:
-      <DateTimePicker @onChange={{fn this.changeDate 'geplandeStart'}} @value={{this.geplandeStart}}/>
-    </div>
-    <div>
-      Locatie[@TN]: <Input type='text' @value={{mut this.opLocatie}} />
-    </div>
-    <div>
-      Started at[@TN]:
-        <DateTimePicker @onChange={{fn this.changeDate 'gestartOpTijdstip'}} @value={{this.gestartOpTijdstip}}/>
-    </div>
-    <div>
-      Finished at[@TN]:
-        <DateTimePicker @onChange={{fn this.changeDate 'geeindigdOpTijdstip'}} @value={{this.geeindigdOpTijdstip}}/>
-    </div>
-    <div>
-      <PowerSelect
-        @placeholder="Selecteer bestuursorgaan"
-        @selected={{this.bestuursorgaan}}
-        @options={{this.bestuursorgaanOptions}}
-        @onchange={{this.changeSelect}} as |bestuursorgaan|>
-          {{bestuursorgaan.isTijdsspecialisatieVan.naam}}
-          (periode: {{moment-format bestuursorgaan.bindingStart "MM/DD/YYYY"}} -
-          {{#if bestuursorgaan.bindingEinde}}{{moment-format bestuursorgaan.bindingEinde "MM/DD/YYYY"}}{{else}}nvt.{{/if}})
-      </PowerSelect>
-    </div>
-    <ParticipationList
-      @bestuursorgaan={{this.bestuursorgaan}}
-      @aanwezigenBijStart={{this.aanwezigenBijStart}}
-      @voorzitter={{this.voorzitter}}
-      @secretaris={{this.secretaris}}
-      @onSave={{this.saveParticipationList}}/>
+<AuHeading @skin="1" class="">Zitting</AuHeading>
 
-     <div>
-      Agendapunten[@TN]:
-      <ManageAgendaZitting @zitting={{@zitting}}> </ManageAgendaZitting>
-    </div>
-  <WuButton @onClick={{this.save}}>Save[@TN]</WuButton>
-</div>
+<AuHeading @skin="2" class="">Algemene informatie over de zitting</AuHeading>
+<Zitting::ManageZittingsdata
+  @zitting={{@zitting}}
+  @onCreate={{@onCreateMeeting}}
+/>
+
+{{#if this.zitting.bestuursorgaan}}
+  <AuHeading @skin="2" class="">Aanwezigen bij start van de zitting</AuHeading>
+  <ParticipationList
+    @bestuursorgaan={{this.zitting.bestuursorgaan}}
+    @aanwezigenBijStart={{this.aanwezigenBijStart}}
+    @voorzitter={{this.voorzitter}}
+    @secretaris={{this.secretaris}}
+    @onSave={{this.saveParticipationList}}
+  />
+
+  <AuHeading @skin="2" class="">Agenda van de zitting </AuHeading>
+  <ManageAgendaZitting @zitting={{@zitting}}> </ManageAgendaZitting>
+{{else}}
+  Alvorens u verder kan gaan, dient de zitting eerst informatie te hebben over het bestuursorgaan.
+{{/if}}
