+++ app/components/zitting/manage-zittingsdata.hbs
@@ -1,0 +1,66 @@
+<div {{did-insert this.fetchBestuursorgaan}}>
+  {{#if this.showModal}}
+    {{#wu-modal
+        title="Beheer algemene informatie over de zitting"
+        dialog-class="modal-dialog--wide modal-dialog--sectioned"
+        onClose=this.toggleModal
+        as |m|
+    }}
+      {{#m.content isSectioned=true}}
+        <div class="u-spacer">
+          <div class="grid">
+            <div class="col--6-12">
+              <label class="form__label">
+                Gehouden door:
+                <PowerSelect
+                  @placeholder="Selecteer bestuursorgaan"
+                  @selected={{this.bestuursorgaan}}
+                  @options={{this.bestuursorgaanOptions}}
+                  @onchange={{this.changeSelect}} as |bestuursorgaan|>
+                    {{bestuursorgaan.isTijdsspecialisatieVan.naam}}
+                    (periode: {{moment-format bestuursorgaan.bindingStart "MM/DD/YYYY"}} -
+                    {{#if bestuursorgaan.bindingEinde}}{{moment-format bestuursorgaan.bindingEinde "MM/DD/YYYY"}}{{else}}nvt.{{/if}})
+                </PowerSelect>
+              </label>
+            </div>
+          </div>
+        </div>
+        <div>
+        Geplande start:
+        <DateTimePicker @onChange={{fn this.changeDate 'geplandeStart'}} @value={{this.geplandeStart}}/>
+      </div>
+      <div>
+        Op locatie: <Input type='text' @value={{mut this.opLocatie}} />
+      </div>
+      <div>
+        Gestart op:
+          <DateTimePicker @onChange={{fn this.changeDate 'gestartOpTijdstip'}} @value={{this.gestartOpTijdstip}}/>
+      </div>
+      <div>
+        Beëindigd op:
+          <DateTimePicker @onChange={{fn this.changeDate 'geeindigdOpTijdstip'}} @value={{this.geeindigdOpTijdstip}}/>
+      </div>
+      {{/m.content}}
+      <div class="modal-dialog__footer">
+        {{#wu-button-group}}
+            <WuButton @onClick={{this.saveZittingsData}}>Voeg zittingsdata in</WuButton>
+            <WuButton @onClick={{this.toggleModal}} class="button button--alt">Terug</WuButton>
+        {{/wu-button-group}}
+      </div>
+    {{/wu-modal}}
+  {{else}}
+      <div>
+        Geplande start: {{moment-format  this.geplandeStart "DD/MM/YYYY HH:mm"}}
+      </div>
+      <div>
+        Op locatie: {{this.opLocatie}}
+      </div>
+      <div>
+        Gestart op: {{moment-format this.gestartOpTijdstip "DD/MM/YYYY HH:mm"}}
+      </div>
+      <div>
+        Beëindigd op: {{moment-format this.geeindigdOpTijdstip "DD/MM/YYYY HH:mm"}}
+      </div>
+      <WuButton @onClick={{this.toggleModal}}>Beheer algemene zittingsdata</WuButton>
+  {{/if}}
+</div>
