+++ app/styles/_shame.scss
@@ -110,7 +110,7 @@ li.link-list__item > button.button {
 /* temporary fix for besluittypes dropdown */
 
 .ember-basic-dropdown-content.ember-power-select-dropdown.ember-basic-dropdown-content--below {
-  z-index: 1006;
+  z-index: 1010;
 }
 
 /* temporary demo fixes for fracties */
