+++ app/templates/inbox.hbs
@@ -38,8 +38,8 @@
             </li>
             <li class="au-c-list-navigation__item">
               <AuNavigationLink @linkRoute="inbox.meetings">
-                <AuIcon @icon="bin" @alignment="left" />
-                Zittings[@TN]
+                <AuIcon @icon="documents" @alignment="left" />
+                Zittingen
               </AuNavigationLink>
             </li>
           </ul>
