+++ app/templates/meetings/edit.hbs
@@ -5,4 +5,5 @@
     </AcmidmLogout>
   </AuDropdown>
 </AuMainHeader>
-<MeetingForm @save={{action "save"}} @zitting={{@model}}/>
+<MeetingForm @zitting={{@model}}/>
+<AuLink @linkRoute="inbox.meetings" @skin="secondary">Terug</AuLink>
